fix type casting and two conditions always being true or false (#6602)

+++ src/Serenity.Net.Core/Caching/TwoLevelCacheExtensions.cs
@@ -26,7 +26,7 @@ namespace Serenity
             int i2 = BitConverter.ToInt32(raw, 4);
             int i3 = BitConverter.ToInt32(raw, 8);
             int i4 = BitConverter.ToInt32(raw, 12);
-            long val = i1 + i2 + i3 + i4;
+            long val = (long)i1 + i2 + i3 + i4;
             while (val > int.MaxValue)
                 val -= int.MaxValue;
             return (int)val;
